move make_meta_dict into a separate file
